modified:   scripts/temples.js 	modified:   styles/temples.css 	modified:   temples.html

+++ scripts/temples.js
@@ -4,7 +4,7 @@ const today = new Date();
 short.innerHTML = `Last modified: <span class="highlight">${new Intl.DateTimeFormat(
 	"en-US",
 	{
-		dateStyle: "short"
+		dateStyle: "long"
 	}
 ).format(today)}</span>`;
